+++ src/main/java/com/spring/webflux/todo/exception/InvalidTodoException.java
@@ -7,15 +7,15 @@ import org.springframework.http.HttpStatus;
 @Getter
 @EqualsAndHashCode(callSuper = true)
 public class InvalidTodoException extends TodoRuntimeException {
-  private final Long todoId;
+  private final Integer todoId;
   private static final HttpStatus httpStatus = HttpStatus.NOT_FOUND;
 
-  public InvalidTodoException(Long todoId, String message) {
+  public InvalidTodoException(Integer todoId, String message) {
     super(httpStatus, message);
     this.todoId = todoId;
   }
 
-  public InvalidTodoException(Long todoId) {
+  public InvalidTodoException(Integer todoId) {
     super(httpStatus);
     this.todoId = todoId;
   }
